# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WebForge PaaS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  telegramId: number;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  tier: string;
  dailyActionsCounter: number;
  projectCount: number;
  isActive: boolean;
  /** @nullable */
  createdAt?: string | null;
}

export interface PlatformStats {
  totalUsers: number;
  totalProjects: number;
  starterUsers: number;
  proUsers: number;
  eliteUsers: number;
}

export interface Project {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: string;
  /** @nullable */
  port?: number | null;
  /** @nullable */
  workDir?: string | null;
  /** @nullable */
  techStack?: string | null;
  isHosted: boolean;
  /** @nullable */
  createdAt?: string | null;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ProjectInput {
  userId: number;
  name: string;
  description?: string;
  techStack?: string;
}

export interface PersonaUpdate {
  name?: string;
  systemPrompt: string;
  tone?: string;
  instructions?: string;
}

export interface Payment {
  id: number;
  userId: number;
  reference: string;
  tier: string;
  amountNgn: number;
  status: string;
  /** @nullable */
  receiptUrl?: string | null;
  /** @nullable */
  createdAt?: string | null;
}

export interface AiGenerateInput {
  userId: number;
  /** planning | coding | fixing | chat | image | ui */
  taskType: string;
  prompt: string;
  /** @nullable */
  projectId?: number | null;
}

export interface AiGenerateResult {
  content: string;
  model: string;
  inputTokens: number;
  outputTokens: number;
  costUsd: number;
}

export interface DeleteResult {
  success: boolean;
}

export type GetMeParams = {
telegramId: number;
};

export type ListProjectsParams = {
userId: number;
};

export type ListPaymentsParams = {
userId: number;
};