# lib/api-zod/src/generated/types/aiGenerateResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WebForge PaaS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiGenerateResult {
  content: string;
  model: string;
  inputTokens: number;
  outputTokens: number;
  costUsd: number;
}