# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WebForge PaaS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Payment {
  id: number;
  userId: number;
  reference: string;
  tier: string;
  amountNgn: number;
  status: string;
  /** @nullable */
  receiptUrl?: string | null;
  /** @nullable */
  createdAt?: string | null;
}