# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WebForge PaaS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  telegramId: number;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  tier: string;
  dailyActionsCounter: number;
  projectCount: number;
  isActive: boolean;
  /** @nullable */
  createdAt?: string | null;
}